Change overall files.. I believe everyone know the changes.
CIB - Everyone.

+++ Gam200Engine/Graphics/TextPlatform.cpp
@@ -25,7 +25,7 @@ namespace Graphics
     {
         BuildNewMeshesIfNeeded();
         std::vector<std::pair<const Vertices*, const Texture*>> result;
-        for (int i = 0; i < vertices.size(); ++i)
+        for (int i = 0; i < static_cast<int>(vertices.size()); ++i)
         {
             if (vertices.at(i).GetVerticesCount() != 0)
             {
